Version 2.0 Altered key bindings for viewing raw diffs from "j" to "u" so as not to interfere with evil-mode key bindings, if they may be in effect.
Fixed a bug that made highlighting differences crash without giving an error if tab characters were located in an unfortunate position in the document.

Also fixed an annoying error that happens every time the latest revision is the first revision to be viewed upon starting the 7z-revisions view.

Added an error that occurs when trying to create an archive from a file, the name of which takes the form of a floating number, in order to enforce the constraint that a file cannot take the name of a floating number.  Eventually, in a later version, this constraint will be removed.

Not yet fixed the bug that causes revisions from the same archive to not be viewable from both MS windows OS and linux OS, interchangeably, although the view diff file function still works interchangeably on both.  In addition, the patch.exe file is still required under windows, the patch shell command still not yet having been replaced by an elisp implementation, which would get rid of the annoying windows 7 pop ups prompting for the allowing of patch.exe to run every time a revision is viewed.