Version 2.3
Fixed a bug that causes an error if the buffer of the main document is already killed, when trying to exit the summary view using the q key.

Fixed a strange bug that, for some reason, prevented the 7z u command from working when running on linux with a dos file encoding.  Well, didnt really fix it, just resorted to using 7z a instead.

Forced read only when opening metadata buffers to prevent accidentally hozing of data.  Anyone fiddling with those files will no doubt know how to disable read-only on the buffer before editing.

Fixed a regression bug that prevented a revision from being viewable if its file was initially created under windows using a DOS file encoding and then later appended to under linux.  It's really fixed this time.  I promise.