Version 3.0
Added a function to highlight all revision numbers that have notes associated with them on the summary list page, bound to key "H".
Fixed various bugs re: viewing revisions, still buggy tho.
Added a function to go to the revision for which a selected region was last modified, a selected region taken from either the original document or a viewed revision.
Added an interactive function called 7zr-refresh-all-bookmarks which updates all the bookmarks in the bookmark-default-file, which point to buffers that have associated .7z archives.  It does this for each bookmark by finding the past revision that matches the bookmarks location and context pattern, then calculating its new line number based on how many revisions added lines before it.