Version 3.2
This program once agains stipulates the installation of programs such as FCIV.exe and gawk.exe when run under windows, reversing the version 3.1 decision.  Program execution is just faster that way.

Added a function to regenerate all hashes from diffs, in the event that the hash table somehow gets hosed.

Also, fixed some bugs with the dired-mode hook, and changed the key binding to C-c z

In this version, lexical-binding has been enabled, but the script is still compatible with emacs 23

There is a particularly tiresome recurring bug re: view revision that causes the 'revision hashes dont match' error when stepping upward in the revision list, but not downwards, which hasnt yet been fixed.